prepare the list of articles
I am still not sure how to organize these posts. Some are specific to an attraction (e.g. castle), but others are collection (restaurants, hiking, biking). I'd love to just reference from a collection to specific article if that is possible to keep one source of truth. I have a feeling that won't be easy. Anyway, keeping the content fresh and consistent is my no 1 priority. There is no point in having yet another guide without it.